common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "opentelemetry-distro>=0.55b1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "pika>=1.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]